Moved getRelationship() from just Role model to trait implemented in all Voyager models
