+++ src/Models/Permission.php
@@ -3,9 +3,12 @@
 namespace TCG\Voyager\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use TCG\Voyager\Traits\HasRelationships;
 
 class Permission extends Model
 {
+    use HasRelationships;
+
     protected $guarded = [];
 
     public function roles()
