+++ src/Models/Role.php
@@ -4,11 +4,13 @@ namespace TCG\Voyager\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use TCG\Voyager\Facades\Voyager;
+use TCG\Voyager\Traits\HasRelationships;
 
 class Role extends Model
 {
+    use HasRelationships;
+
     protected $guarded = [];
-    protected static $relationships = [];
 
     public function users()
     {
@@ -18,14 +20,5 @@ class Role extends Model
     public function permissions()
     {
         return $this->belongsToMany(Voyager::modelClass('Permission'));
-    }
-
-    public static function getRelationship($id)
-    {
-        if (!isset(self::$relationships[$id])) {
-            self::$relationships[$id] = self::find($id);
-        }
-
-        return self::$relationships[$id];
     }
 }
