+++ src/Models/User.php
@@ -5,11 +5,13 @@ namespace TCG\Voyager\Models;
 use Carbon\Carbon;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use TCG\Voyager\Contracts\User as UserContract;
+use TCG\Voyager\Traits\HasRelationships;
 use TCG\Voyager\Traits\VoyagerUser;
 
 class User extends Authenticatable implements UserContract
 {
-    use VoyagerUser;
+    use VoyagerUser,
+        HasRelationships;
 
     protected $guarded = [];
